Extract base of czech and english "dates.py" to base "dates.py"
